pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cleanbox"
version = "1.0.0"
description = "Windows disk cleanup utility with system tray integration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "CleanBox Team"}
]
keywords = ["disk", "cleanup", "windows", "utility", "storage"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    "PyQt6>=6.6.0",
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
    "win11toast>=0.34.0",
    "psutil>=5.9.0",
    "winshell>=0.6",
    "pywin32>=306",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-qt>=4.0.0",
    "pytest-cov>=4.0.0",
    "coverage>=7.0.0",
    "flake8>=6.0.0",
    "bandit>=1.7.0",
    "radon>=5.0.0",
    "autopep8>=2.0.0",
]

[project.scripts]
cleanbox = "main:main"

[project.gui-scripts]
cleanbox-gui = "main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = []

[tool.flake8]
max-line-length = 120
exclude = [".venv", "__pycache__", "tests"]
